renderer: Move draw to its own file
